Add verifyRoutingPolicyReferences to finalizeConfiguration
When assertions are enabled, crashes if a CallExpr or CallStatement
refers to a nonexistent routing policy. This would have caught the
issue fixed by commit 1bc0b0dd77 where PsFromPolicyStatement was
creating a CallExpr for a nonexistent policy.

---
Prompt:
```
I'd like to add a verifyRoutingPolicyReferences(c) function in finalizeConfiguration. It is styled after verifyCommunityStructures, but for RoutingPolicies. When assertions are enabled, it crashes if a CallExpr or CallStatement is present that refers to a non-existing routing policy. Crucially, it would have caught the Juniper PsFromPolicyStatement issue fixed by the most recent commit.
```

commit-id:1a492b98