IOS XR and Juniper: omit CallExpr/CallStatement when policy doesn't exist
Similar to batfish/batfish#9553. When a routing policy references another
policy that doesn't exist, omit the CallExpr or CallStatement instead of
creating a reference to a nonexistent routing policy.

Fixes:
- IOS XR RoutePolicyApplyStatement: check before creating CallStatement
- IOS XR RoutePolicyBooleanApply: return FALSE instead of CallExpr
- Juniper rib-group import-policy: filter out undefined policies
- Juniper resolution rib import-policy: filter out undefined policies

---
Prompt:
```
Let's also fix the tests failing in `bazel test //tests/parsing-tests:ref_tests --test_output=streamed`. Note that you probably want to throw in a grep for org.batfish to avoid a huge amount of output and focus only on the failing stack traces.
```

commit-id:fb2ca65a